Extract agent message consumption allocation
